chore: refactoring data models
