+++ src/components/SkillMatrix.tsx
@@ -1,7 +1,7 @@
 import { $, component$, useContext, useSignal, useTask$ } from '@builder.io/qwik';
 import { AppContext } from '../app';
 import { tt } from '../locale/labels';
-import { Skill } from '../models/types';
+import { Skill } from '../models/Month';
 import { SkillRow } from './SkillRow';
 import { getSkillMatrixMine, pathSkillMatrixMine } from '../services/skillMatrix';
