+++ src/models/Customer.ts
@@ -1,0 +1,1 @@
+export type Customer = string;
