+++ src/models/Month.ts
@@ -1,0 +1,6 @@
+export type Month = {
+	month_year: string;
+	confirmedEffort: number;
+	tentativeEffort: number;
+	notes: string;
+};
