+++ src/models/OpenAIResponse.ts
@@ -1,0 +1,1 @@
+export type OpenAIResponse = { message: string };
