+++ src/services/configuration.ts
@@ -1,4 +1,4 @@
-import { Configuration } from '../models/types';
+import { Configuration } from '../models/Month';
 import { getHttpResponse } from '../network/httpRequest';
 
 export const getConfiguration = async (): Promise<Configuration> =>
