+++ src/services/skillMatrix.ts
@@ -1,4 +1,4 @@
-import { Skill, SkillMatrix } from '../models/types';
+import { Skill, SkillMatrix } from '../models/Month';
 import { getHttpResponse, checkHttpResponseStatus } from '../network/httpRequest';
 
 export const getSkillMatrixMine = async (): Promise<Skill[]> =>
